# packages/api-contract/openapi/footalert.v1.yaml
openapi: 3.1.0
info:
  title: FootAlert BFF API
  version: 1.0.0
  description: |
    Public contract consumed by FootAlert clients (mobile, web, desktop).
    This spec intentionally uses broad payload schemas because upstream API-Football
    structures can evolve, while preserving endpoint/query/path contracts.
servers:
  - url: http://localhost:3001
    description: Local BFF server
  - url: https://staging-api.footalert.example.com
    description: Staging BFF server
paths:
...
  /v1/telemetry/events:
    post:
      summary: Push telemetry event (attested mobile session)
      operationId: sendTelemetryEvent
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryEventPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/telemetry/events/batch:
    post:
      summary: Push telemetry event batch (attested mobile session)
      operationId: sendTelemetryEventBatch
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/TelemetryEventPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/telemetry/errors:
    post:
      summary: Push telemetry error (attested mobile session)
      operationId: sendTelemetryError
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryErrorPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/telemetry/errors/batch:
    post:
      summary: Push telemetry error batch (attested mobile session)
      operationId: sendTelemetryErrorBatch
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/TelemetryErrorPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/telemetry/breadcrumbs:
    post:
      summary: Push telemetry breadcrumb (attested mobile session)
      operationId: sendTelemetryBreadcrumb
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryBreadcrumbPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
  /v1/telemetry/breadcrumbs/batch:
    post:
      summary: Push telemetry breadcrumb batch (attested mobile session)
      operationId: sendTelemetryBreadcrumbBatch
      security:
        - MobileSessionBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/TelemetryBreadcrumbPayload'
      responses:
        '200':
          $ref: '#/components/responses/AcceptedTelemetryResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
    x-platform-scope:
      - mobile
components:
  securitySchemes:
    MobileSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    AcceptedTelemetryResponse:
      description: Telemetry payload accepted
      content:
        application/json:
          schema:
            type: object
            required: [ status, type ]
            properties:
              status:
                type: string
                enum: [ accepted ]
              type:
                type: string
                enum: [ event, event_batch, error, error_batch, breadcrumb, breadcrumb_batch ]
              count:
                type: integer
                minimum: 1
    ProblemResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemResponse'
  schemas:
...
    ProblemResponse:
      type: object
      required: [ error, message ]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: 'null'
...
    TelemetryScalar:
      oneOf:
        - type: string
        - type: number
        - type: boolean
        - type: 'null'
    TelemetryAttributes:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/TelemetryScalar'
    TelemetryEventPayload:
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        attributes:
          $ref: '#/components/schemas/TelemetryAttributes'
        userContext:
          $ref: '#/components/schemas/TelemetryAttributes'
        timestamp:
          type: string
          format: date-time
    TelemetryErrorContext:
      type: object
      additionalProperties: false
      properties:
        feature:
          type: string
          maxLength: 120
        status:
          type: integer
          minimum: 100
          maximum: 599
        url:
          type: string
          maxLength: 2048
        method:
          type: string
          maxLength: 16
        details:
          $ref: '#/components/schemas/TelemetryAttributes'
    TelemetryErrorPayload:
      type: object
      additionalProperties: false
      required: [ name, message ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 240
        message:
          type: string
          minLength: 1
          maxLength: 2000
        stack:
          type: string
          maxLength: 20000
        context:
          $ref: '#/components/schemas/TelemetryErrorContext'
        userContext:
          $ref: '#/components/schemas/TelemetryAttributes'
        timestamp:
          type: string
          format: date-time
    TelemetryBreadcrumbPayload:
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        attributes:
          $ref: '#/components/schemas/TelemetryAttributes'
        userContext:
          $ref: '#/components/schemas/TelemetryAttributes'
        timestamp:
          type: string
          format: date-time